pyproject.toml
[project]
name = "node-gallery-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "HansenBurger", email = "1035376945@qq.com"},
]
dependencies = ["fastapi>=0.115.12", "sqlmodel>=0.0.24", "uvicorn>=0.34.2", "psycopg2-binary>=2.9.10", "python-jose[cryptography]>=3.4.0", "passlib[bcrypt]>=1.7.4", "python-multipart>=0.0.20"]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
